Split Autocomplete component to smaller parts
