+++ components/materialUI/Autocomplete/Menu.js
@@ -1,0 +1,27 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import Paper from '@material-ui/core/Paper';
+
+export default function Menu(props) {
+  return (
+    <Paper
+      square
+      className={props.selectProps.classes.paper}
+      {...props.innerProps}
+    >
+      {props.children}
+    </Paper>
+  );
+}
+
+Menu.propTypes = {
+  /**
+   * The children to be rendered.
+   */
+  children: PropTypes.element.isRequired,
+  /**
+   * Props to be passed to the menu wrapper.
+   */
+  innerProps: PropTypes.object.isRequired,
+  selectProps: PropTypes.object.isRequired,
+};
