+++ components/materialUI/Autocomplete/Option.js
@@ -1,0 +1,55 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import MenuItem from '@material-ui/core/MenuItem';
+
+export default function Option(props) {
+  return (
+    <MenuItem
+      ref={props.innerRef}
+      selected={props.isFocused}
+      component="div"
+      style={{
+        fontWeight: props.isSelected ? 500 : 400,
+      }}
+      {...props.innerProps}
+    >
+      {props.children}
+    </MenuItem>
+  );
+}
+
+Option.propTypes = {
+  /**
+   * The children to be rendered.
+   */
+  children: PropTypes.node,
+  /**
+   * props passed to the wrapping element for the group.
+   */
+  innerProps: PropTypes.shape({
+    id: PropTypes.string.isRequired,
+    key: PropTypes.string,
+    onClick: PropTypes.func.isRequired,
+    onMouseMove: PropTypes.func.isRequired,
+    onMouseOver: PropTypes.func.isRequired,
+    tabIndex: PropTypes.number.isRequired,
+  }).isRequired,
+  /**
+   * Inner ref to DOM Node
+   */
+  innerRef: PropTypes.oneOfType([
+    PropTypes.oneOf([null]),
+    PropTypes.func,
+    PropTypes.shape({
+      current: PropTypes.any.isRequired,
+    }),
+  ]),
+  /**
+   * Whether the option is focused.
+   */
+  isFocused: PropTypes.bool.isRequired,
+  /**
+   * Whether the option is selected.
+   */
+  isSelected: PropTypes.bool.isRequired,
+};
