+++ components/materialUI/Autocomplete/Placeholder.js
@@ -1,0 +1,28 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import Typography from '@material-ui/core/Typography';
+
+export default function Placeholder(props) {
+  const { selectProps, innerProps = {}, children } = props;
+  return (
+    <Typography
+      color="textSecondary"
+      className={selectProps.classes.placeholder}
+      {...innerProps}
+    >
+      {children}
+    </Typography>
+  );
+}
+
+Placeholder.propTypes = {
+  /**
+   * The children to be rendered.
+   */
+  children: PropTypes.node,
+  /**
+   * props passed to the wrapping element for the group.
+   */
+  innerProps: PropTypes.object,
+  selectProps: PropTypes.object.isRequired,
+};
