+++ components/materialUI/Autocomplete/SingleValue.js
@@ -1,0 +1,26 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import Typography from '@material-ui/core/Typography';
+
+export default function SingleValue(props) {
+  return (
+    <Typography
+      className={props.selectProps.classes.singleValue}
+      {...props.innerProps}
+    >
+      {props.children}
+    </Typography>
+  );
+}
+
+SingleValue.propTypes = {
+  /**
+   * The children to be rendered.
+   */
+  children: PropTypes.node,
+  /**
+   * Props passed to the wrapping element for the group.
+   */
+  innerProps: PropTypes.any,
+  selectProps: PropTypes.object.isRequired,
+};
