+++ components/materialUI/Autocomplete/ValueContainer.js
@@ -1,0 +1,18 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+export default function ValueContainer(props) {
+  return (
+    <div className={props.selectProps.classes.valueContainer}>
+      {props.children}
+    </div>
+  );
+}
+
+ValueContainer.propTypes = {
+  /**
+   * The children to be rendered.
+   */
+  children: PropTypes.node,
+  selectProps: PropTypes.object.isRequired,
+};
